move view metadata over to separate folder
